# Speed up bulk get/read