--- a/lib/content/read.js
+++ b/lib/content/read.js
@@ -4,11 +4,30 @@ const Promise = require('bluebird')
 
 const checksumStream = require('checksum-stream')
 const contentPath = require('./path')
+const crypto = require('crypto')
 const fs = require('graceful-fs')
 const pipeline = require('mississippi').pipeline
 
 Promise.promisifyAll(fs)
 
+module.exports = read
+function read (cache, address, opts) {
+  opts = opts || {}
+  const algo = opts.hashAlgorithm || 'sha512'
+  const cpath = contentPath(cache, address, algo)
+  return fs.readFileAsync(cpath, null).then(data => {
+    const digest = crypto.createHash(algo).update(data).digest('hex')
+    if (typeof opts.size === 'number' && opts.size !== data.length) {
+      throw sizeError(opts.size, data.length)
+    } else if (digest !== address) {
+      throw checksumError(address, digest)
+    } else {
+      return data
+    }
+  })
+}
+
+module.exports.stream = readStream
 module.exports.readStream = readStream
 function readStream (cache, address, opts) {
   opts = opts || {}
@@ -36,4 +55,20 @@ function hasContent (cache, address, algorithm) {
       throw err
     }
   })
+}
+
+function sizeError (expected, found) {
+  var err = new Error('stream data size mismatch')
+  err.expected = expected
+  err.found = found
+  err.code = 'EBADSIZE'
+  return err
+}
+
+function checksumError (expected, found) {
+  var err = new Error('checksum failed')
+  err.code = 'EBADCHECKSUM'
+  err.expected = expected
+  err.found = found
+  return err
 }