--- a/test/util/bufferise.js
+++ b/test/util/bufferise.js
@@ -1,0 +1,8 @@
+'use strict'
+
+module.exports = bufferise
+function bufferise (string) {
+  return Buffer.from
+  ? Buffer.from(string, 'utf8')
+  : new Buffer(string, 'utf8')
+}