This does something similar to #59 but for `content.read()` and `get()`: There are now bulk operations that read everything into memory in one shot, which can be used for known-small data.

The boost is pretty significant:

### Benchmarks

(NOTE: I don't wanna bother doing cross-version comparisons, but it's probably helpful to point out that `get()` was a tinge slower than `get.stream()` before, and `get.stream()` is unchanged with this PR)

```
================================================
     content.read()
------------------------------------------------
  9660 ops/s @ ~0.104ms/op (±1.23%)
  Sampled 70 in 6.29s.
================================================
     content.read.stream() small data
------------------------------------------------
  3079 ops/s @ ~0.325ms/op (±2.87%)
  Sampled 58 in 6.02s.
================================================
     content.read.stream() big data
------------------------------------------------
  155 ops/s @ ~6.457ms/op (±3.25%)
  Sampled 65 in 5.87s.
================================================
     get.byDigest()
------------------------------------------------
  8278 ops/s @ ~0.121ms/op (±1.98%)
  Sampled 69 in 6.80s.
================================================
     get.byDigest() memoized
------------------------------------------------
  97456 ops/s @ ~0.010ms/op (±1.78%)
  Sampled 79 in 6.03s.
================================================
     get.stream.byDigest() small data
------------------------------------------------
  2898 ops/s @ ~0.345ms/op (±3.06%)
  Sampled 59 in 6.10s.
================================================
     get.stream() big data
------------------------------------------------
  152 ops/s @ ~6.597ms/op (±2.51%)
  Sampled 64 in 5.96s.
================================================
```